PART2/Users/requirements.txt
flask
flask-restful
mysql-connector
requests